move group odrdering to utils file
